iot4-nfc-tag-identification/requirements.txt
adafruit-circuitpython-pn532
lgpio
pycardano
python-dotenv
requests
